Fix overflow, check memory leak
Changed int to long long

+++ Offline 08 (Greedy)/main.cpp
@@ -1,5 +1,5 @@
 #include<iostream>
-#include<algorithm>
+//#include<algorithm>
 #include "quicksort_descending.h"
 
 using namespace std;
@@ -7,11 +7,11 @@ using namespace std;
 int main() {
     int n, k;
     cin >> n >> k;
-    int *plants = new int[n];
+    long long *plants = new long long[n];
     for (int i = 0; i < n; i++) cin >> plants[i];
-//    sort(plants, plants+n, greater<int>());
+//    sort(plants, plants+n, greater<long long>());
     quicksort_descending(plants, 0, n-1);
-    int cost = 0;
+    long long cost = 0;
     for (int i = 0; i < n; i++) {
         cost += (i/k + 1) * plants[i];
     }
@@ -20,7 +20,7 @@ int main() {
     return 0;
 }
 
-// Proof of correctness
+// Proof of correctness:
 // Let a_1, a_2, a_3, ..., a_n be the array of plant cost, sorted in descending order of values i.e. a_i >= a_i+1
 // and b_1, b_2, b_3, ..., b_n be the factor by which each plant price is multiplied, where b_i <= b_i+1
 // Our claim is, minimum cost will be a_1 * b_1 + a_2 * b_2 + ... + a_n * b_n = c
@@ -42,5 +42,3 @@ int main() {
 // b_1 = b_2 = b_3 = ... = b_k = 0
 // b_k+1 = b_k+2 = ...= b_2k = 1 and so on up to n
 // which gives us the optimal solution i.e. the minimum cost (QED)
-
-
